Speedup arrayposn update when len(servers) << len(local_domains)
There is a use-case that has lots of local domains and a few servers to
forward to. It might be ad-blocking in pi-hole style or DNS-based policy
routing similar to antizapret and zaborona VPNs.

The patch gives 2.5x speedup to arrayposn update step of build_server_array()
on 100,000 domains dataset on SNR hardware: 50ms -> 20ms.

That's 10% of overall build_server_array() runtime for this test-case
bringing it down to 313ms on the aforementioned router.