Use a bit-field to quick-skip lookups with no results
It gives no measurable performance improvements as it's effective for
blocklists without TLD's, so it skips wormbsearch(com) for
lookup_domain(example.com), but these TLD's are probably in CPU dcache
anyway, so the performance benefit is close-to-zero.

It speedups pathological case of a random full-length one-byte-label
domain like ('$(( $RANDOM % 10 )).' * (254 // 2)): 634us -> 208us.